+++ cvat-ui/src/reducers/annotation-reducer.ts
@@ -51,6 +51,7 @@ const defaultState: AnnotationState = {
             number: 0,
             filename: '',
             data: null,
+            hasRelatedContext: false,
             fetching: false,
             delay: 0,
             changeTime: null,
@@ -58,8 +59,8 @@ const defaultState: AnnotationState = {
         playing: false,
         frameAngles: [],
         contextImage: {
-            loaded: false,
-            data: '',
+            fetching: false,
+            data: null,
             hidden: false,
         },
     },
@@ -145,6 +146,7 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
                 openTime,
                 frameNumber: number,
                 frameFilename: filename,
+                frameHasRelatedContext,
                 colors,
                 filters,
                 frameData: data,
@@ -189,6 +191,7 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
                     frame: {
                         ...state.player.frame,
                         filename,
+                        hasRelatedContext: frameHasRelatedContext,
                         number,
                         data,
                     },
@@ -226,11 +229,6 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
                         ...state.player.frame,
                         fetching: false,
                     },
-                    contextImage: {
-                        loaded: false,
-                        data: '',
-                        hidden: state.player.contextImage.hidden,
-                    },
                 },
             };
         }
@@ -252,7 +250,16 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
         }
         case AnnotationActionTypes.CHANGE_FRAME_SUCCESS: {
             const {
-                number, data, filename, states, minZ, maxZ, curZ, delay, changeTime,
+                number,
+                data,
+                filename,
+                hasRelatedContext,
+                states,
+                minZ,
+                maxZ,
+                curZ,
+                delay,
+                changeTime,
             } = action.payload;
 
             const activatedStateID = states
@@ -268,6 +275,7 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
                     frame: {
                         data,
                         filename,
+                        hasRelatedContext,
                         number,
                         fetching: false,
                         changeTime,
@@ -275,7 +283,7 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
                     },
                     contextImage: {
                         ...state.player.contextImage,
-                        loaded: false,
+                        data: null,
                     },
                 },
                 annotations: {
@@ -1170,30 +1178,52 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
         }
         case AnnotationActionTypes.HIDE_SHOW_CONTEXT_IMAGE: {
             const { hidden } = action.payload;
-            const { loaded, data } = state.player.contextImage;
             return {
                 ...state,
                 player: {
                     ...state.player,
                     contextImage: {
-                        loaded,
-                        data,
+                        ...state.player.contextImage,
                         hidden,
                     },
                 },
             };
         }
         case AnnotationActionTypes.GET_CONTEXT_IMAGE: {
-            const { context, loaded } = action.payload;
+            return {
+                ...state,
+                player: {
+                    ...state.player,
+                    contextImage: {
+                        ...state.player.contextImage,
+                        fetching: true,
+                    },
+                },
+            };
+        }
+        case AnnotationActionTypes.GET_CONTEXT_IMAGE_SUCCESS: {
+            const { contextImageData } = action.payload;
 
             return {
                 ...state,
                 player: {
                     ...state.player,
                     contextImage: {
-                        loaded,
-                        data: context,
-                        hidden: state.player.contextImage.hidden,
+                        ...state.player.contextImage,
+                        fetching: false,
+                        data: contextImageData,
+                    },
+                },
+            };
+        }
+        case AnnotationActionTypes.GET_CONTEXT_IMAGE_FAILED: {
+            return {
+                ...state,
+                player: {
+                    ...state.player,
+                    contextImage: {
+                        ...state.player.contextImage,
+                        fetching: false,
                     },
                 },
             };