+++ cvat-ui/src/components/annotation-page/canvas/canvas-wrapper.tsx
@@ -18,6 +18,7 @@ import getCore from 'cvat-core-wrapper';
 import consts from 'consts';
 import CVATTooltip from 'components/common/cvat-tooltip';
 import ImageSetupsContent from './image-setups-content';
+import ContextImage from '../standard-workspace/context-image/context-image';
 
 const cvat = getCore();
 
@@ -773,12 +774,12 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
             maxZLayer,
             curZLayer,
             minZLayer,
-            onSwitchZLayer,
-            onAddZLayer,
             keyMap,
             switchableAutomaticBordering,
             automaticBordering,
             onSwitchAutomaticBordering,
+            onSwitchZLayer,
+            onAddZLayer,
         } = this.props;
 
         const preventDefault = (event: KeyboardEvent | undefined): void => {
@@ -816,6 +817,8 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
                         height: '100%',
                     }}
                 />
+
+                <ContextImage />
 
                 <Dropdown trigger='click' placement='topCenter' overlay={<ImageSetupsContent />}>
                     <UpOutlined className='cvat-canvas-image-setups-trigger' />